change model status, and moved files to dto
