+++ car_rent.Server/Controllers/UserController.cs
@@ -1,8 +1,8 @@
 using car_rent_api2.Server.Database;
-using car_rent.Server.Model;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using car_rent.Server.DTOs;
 
 namespace car_rent.Server.Controllers;
