+++ car_rent.Server/DTOs/EditUserModel.cs
@@ -1,4 +1,4 @@
-namespace car_rent.Server.Model;
+namespace car_rent.Server.DTOs;
 
 public class EditUserModel
 {
