+++ car_rent.Server/Model/Car.cs
@@ -6,11 +6,10 @@ public class Car(string brand, string model, int year, string picture)
     public string Model { get; set; } = model;
     public string Brand { get; set; } = brand;
     public int Year { get; set; } = year;
-
+    public string Picture { get; set; } = picture;
+    
     public string GetCarInfo()
     {
         return $"Model: {Model}, Brand: {Brand}, Year: {Year}";
     }
-
-    public string Picture { get; set; } = picture;
 }
