+++ car_rent.Server/Model/Rent.cs
@@ -2,6 +2,12 @@
 
 namespace car_rent_api2.Server.Database
 {
+    public enum RentStatus
+    {
+        Reserved,
+        Rented,
+        Returned
+    }
     public class Rent
     {
         public Guid Rent_ID { get; set; }
@@ -9,7 +15,7 @@ namespace car_rent_api2.Server.Database
         public DateTime Rent_date { get; set; }
         public DateTime Return_date { get; set; }
         public Guid User_ID { get; set; }
-        public string Status { get; set; }
+        public RentStatus Status { get; set; }
         public Guid Company_ID { get; set; }
         public Guid Offer_ID { get; set; }
         // Navigation properties
