+++ car_rent.client/src/pages/history.jsx
@@ -108,7 +108,7 @@ const History = () => {
                             <td>{rental.offer.price.toString() || "N/A"}</td>
                             <td>{rental.offer.car.brand || "N/A"}</td>
                             <td>
-                                {rental.status !== 'returned' ? (
+                                {rental.status !== 'Returned' ? (
                                     <button className="btn btn-return" onClick={returnOffer(rental.rentId_in_company) }>Return</button>
                                 ) : (
                                     rental.status
